discover an OpenHPSDR radio using the IPv4 address
Use a pre-declared IPv4 address to find the radio and connect to it
without broadcast discovery. The user can input the IPv4 in the
discovery window. If the user declared IPv4 address is empty the
discovery is disabled. Moreover, if the IP is already discovered by the
broadcast procedure the static IP is not used.

This is very useful when you use a VPN or you computer is connected to
another network (therefore is leveraging the routing mechanism).
On both cases, you need to pay attention to the firewall rules.

To maintain the interface easy to use, the TCP (aka red pitaya)
and UDP static IP input boxes are merged in a single input.