equalizer menu: Make 'Enable' button react to RXn or TX choice.
From the menu one can set equalizer values for receivers and transmitter. But when selecting as receiver or the transmitter, whether the selected item is actually enabled or not, is not visible.
This patch makes the 'Enable' reflect the state of the selected receiver or transmitter. After having selected a receiver or transmitter and then toggling the 'Enable' this still has the intended effect of enabling/disabling.